[#44] Upgrade Mathics, remove deadsnakes dependency
The deadsnakes PPA has apparently not been updated for compatibility
with Ubuntu Groovy, so the build was failing. I think we only needed
deadsnakes (which provides older Python versions) for Mathics,
which (until the recent new release) had a bug that prevented using it
with Python 3.8. By upgrading Mathics, we remove the need for the old
Python version, which means we don't have to figure out how to work
around the deadsnakes compatibility problem.