Include `-m` flag in `docker run` call to limit memory per container
As per Docker run documentation (https://docs.docker.com/config/containers/resource_constraints/#limit-a-containers-access-to-memory), the `-m` flag must be set in order for `--oom-kill-disable` to have the intended effect. Otherwise, an exhaustion of memory in the container will put pressure on the host's processes. This commit sets the max memory to 200MB assuming the t3.small instance has 2GB total.